# docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.User:
    type: object
  models.UserCurrency:
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 高性能虚拟货币管理系统
  title: Currency Management System API
  version: "1.0"
paths:
  /addCurrencyNum:
    post:
      consumes:
      - application/json
      description: 增加用户的货币数量
      parameters:
      - description: 用户货币信息
        in: body
        name: userCurrency
        required: true
        schema:
          $ref: '#/definitions/models.UserCurrency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 增加货币数量
      tags:
      - 货币管理
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录并返回token
      parameters:
      - description: 登录信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 用户登录
      tags:
      - 用户管理
  /register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 用户注册
      tags:
      - 用户管理
  /subtractCurrencyNum:
    post:
      consumes:
      - application/json
      description: 减少用户的货币数量
      parameters:
      - description: 用户货币信息
        in: body
        name: userCurrency
        required: true
        schema:
          $ref: '#/definitions/models.UserCurrency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 减少货币数量
      tags:
      - 货币管理
  /userCurrency:
    post:
      consumes:
      - application/json
      description: 为用户添加新的货币类型
      parameters:
      - description: 用户货币信息
        in: body
        name: userCurrency
        required: true
        schema:
          $ref: '#/definitions/models.UserCurrency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 添加用户货币
      tags:
      - 货币管理
    put:
      consumes:
      - application/json
      description: 更新用户的货币数量
      parameters:
      - description: 用户货币信息
        in: body
        name: userCurrency
        required: true
        schema:
          $ref: '#/definitions/models.UserCurrency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新用户货币
      tags:
      - 货币管理
  /userCurrency/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定用户的货币信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserCurrency'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取用户货币
      tags:
      - 货币管理
schemes:
- http
- https
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"